Split TileInfo from other parts
