+++ application/src/main/scala/com/azavea/franklin/datamodel/TileMatrixSetLink.scala
@@ -1,0 +1,6 @@
+package com.azavea.franklin.datamodel
+
+import io.circe.generic.JsonCodec
+
+@JsonCodec
+case class TileMatrixSetLink(tileMatrixSet: String, tileMatrixSetURI: String)
